frontend/requirements.txt
streamlit
streamlit-option-menu
requests
python-dotenv
streamlit-chat
python-jose[cryptography]
pydantic
